rename volley module to android
