+++ location-client-android/README.md
@@ -1,11 +1,11 @@
-## Location-Client-Volley
+## Location-Client-Android
 location data consumer on android volley
 
 ### Build a *LocationConfig*
 - build by yourself
 ```Java
 LocationConfig config = new NCULocationConfig();
-config.setServerAddress( "http://127.0.0.1:8080/location-Service" )
+config.setServerAddress( "http://127.0.0.1:8080/location-Service/api/v2" )
 ```
 - build by properties file in resource path
 ```Java
@@ -13,7 +13,7 @@ LocationConfig config = new NCULocationConfig().configure( "settings.properties"
 ```
 ```Java
 #properties in config file
-location.server_address = http://127.0.0.1:8080/location-Service
+location.server_address = http://127.0.0.1:8080/location-Service/api/v2
 ```
 
 ### Build a *LocationClient* using *LocationConfig*
@@ -25,7 +25,7 @@ AsynLocationClient client = new NCULocationClient( config, androidContext )
 ```Java
 class VolleyResponseListener implements ResponseListener<T> {
         @Override
-        void onResponse( Set<T> responses ) {
+        void onResponse( List<T> responses ) {
             //do something on reponses
         }
         @Override
@@ -38,6 +38,7 @@ class VolleyResponseListener implements ResponseListener<T> {
 ### Call methods on *AsynLocationClient* to fetch data from server
 - ``` client.getPlaces( "chineseName", new VolleyResponseListener<Place>() ) ```
 - ``` client.getPlaces( PlaceType.SCENE, new VolleyResponseListener<Place>() ) ```
+- ``` client.getPlaceUnits( "chineseName", new VolleyResponseListener<Unit>() ) ```
 - ``` client.getUnits( "fullName", new VolleyResponseListener<Unit>() ) ```
 - ``` client.getPeople( "chineseName", new VolleyResponseListener<Person>() ) ```
 - ``` client.getWords( "keyword", new VolleyResponseListener<Word>() ) ```
