+++ location-client-android/src/main/java/tw/edu/ncu/cc/location/client/android/NCUAsyncLocationClient.java
@@ -1,4 +1,4 @@
-package tw.edu.ncu.cc.location.client.volley;
+package tw.edu.ncu.cc.location.client.android;
 
 
 import android.content.Context;
