+++ settings.gradle
@@ -1,4 +1,4 @@
 include 'location-server'
 include 'location-client-tool'
-include 'location-client-volley'
+include 'location-client-android'
 include 'location-data'
